+++ src/Hyperwallet.js
@@ -559,6 +559,25 @@ export default class Hyperwallet {
     }
 
     //--------------------------------------
+    // Client Token
+    //--------------------------------------
+
+    /**
+     * Get client token
+     *
+     * @param {string} userToken - The user token
+     * @param {api-callback} callback - The callback for this call
+     *
+     * @throws Will throw an error if userToken is not provided
+     */
+    getClientToken(userToken, callback) {
+        if (!userToken) {
+            throw new Error("userToken is required");
+        }
+        this.client.doPost(`users/${encodeURIComponent(userToken)}/client-token`, {}, {}, callback);
+    }
+
+    //--------------------------------------
     // Paper Checks
     //--------------------------------------